+++ imgurpca/base/electronic.py
@@ -11,17 +11,19 @@ import time
 
 class Electronic(object):
 
+    SECOND = 1
     MINUTE = 60
     HOUR = 3600
     DAY = 86400
     WEEK = 604800
 
     def __init__(self, *args, **kwargs):
-        self._interval = None   # interval in s for auto tasks
+        self._interval = 0      # interval in s for auto tasks
         self._func = None       # task to perform after _interval
         self._args = []         # list of args to be passed to _func
-        self._taskthread = None # thread spawned on Bot().go()
+        self._task = None        # instance of derived Parallel running on _taskthread
         self._until = -1        # time limit on running task
+        self._n = -1            # times to call automated task
 
         for attr in kwargs:
             setattr(self, attr, kwargs[attr])
@@ -63,6 +65,17 @@ class Electronic(object):
         return self
 
 
+    def times(self, n):
+        """
+        specify how many times to perform the task. The automated process ends
+        when time limit in until() is reached, or function in do() is called
+        'n' times, whichever comes first.
+        @param n (int): number of times to perform task.
+        """
+        self._n = n
+        return self
+
+
     def go(self):
         """begin scheduled task. Derives and runs Parallel instance with 1 thread.
         """
@@ -72,12 +85,16 @@ class Electronic(object):
                 stop_flag = common[1]
                 interval = common[2]
                 until = common[3]
-                while not (stop_flag.is_set() or (until>0 and time.time()>=until)):
+                n = common[4]
+                i=0
+                while not (stop_flag.is_set() or (until>0 and time.time()>=until) or (n>0 and i>=n)):
                     function(*pkg)
+                    i+=1
                     time.sleep(interval)
+                stop_flag.clear()   # clear stop flag once loop ends
 
         self._task = Task([self._args], nthreads=1)
-        self._task.common = [self._func, threading.Event(), self._interval, self._until]
+        self._task.common = [self._func, threading.Event(), self._interval, self._until, self._n]
         self._task.start()
 
 
@@ -85,4 +102,13 @@ class Electronic(object):
         """stop the task after it comes out of sleep for next execution"""
         self._task.common[1].set()
         if force:
-            self._taskthread.join(timeout=0)
+            self._task.threads[0].join(timeout=0)
+        else:
+            self._task.threads[0].join()
+
+        self._until = -1    # reset all parameters
+        self._n = -1
+        self._args = []
+        self._interval = 0
+        self._func = None
+        self._task = None