+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### CHANGELOG for v2.2.1
+
+* Added `times()` as frequency limit for Bot automated tasks.
+* Fixed possible bugs with `Markov` class.
+* Added `examples\obfuscate.py` as example script.
+
 ### CHANGELOG for v2.2.0
 
 * Added decision trees to `BaseLearner` represented by `base.DTree`.